handle new tail, store latest version of each key
sending the new path to the tail to all nodes when the tail changes. The
nodes connect to the new tail.

Storing the latest committed version of each key. The tail node will use this to
lookup the latest committed version for a key and send the version to
the Nodes that ask for it.