add transport pkg
Separating the transport stuff into a new package and making the node
and coordinator pkg's naive to the transport method used. This will make
it easier to swap net/rpc for other transport mechanisms. This also
makes testing a lot easier because the coordinator and nodes can be
hooked up directly to each other during tests.