node: backfill latest on start, fix neighbor conn
On startup, nodes need to backfill n.latest with the latest committed
versions they already have in their store.

Bug Fixes:

Use the correct Node client when setting the neighbor. This was causing
the rpc client for the predecessor to be nil.

During propagation, while committing, if the key doesn't exist, write it
first and don't return the store.ErrNotFound error. This was causing the
node to report an unknown key during store.Commit even though everything
was working as intended.