store, transport, node: change signature for PropagationRequest
The current propagation requests include every version of every
committed or dirty item, but all the system needs in order to find the
items the node needs is the latest version of each key. This change
replaces the map[string][]uint64 signatures everywhere with
map[string]uint64. From now on, the propagation messages will only
include the highest version for each key.